veh-dam-detect-2/requirement.txt.txt
torch
torchvision
fastapi
uvicorn
streamlit
opencv-python
pillow